backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pypdf
pymupdf
pillow
aiofiles
janome